enrichment-service/requirements.txt
transformers==4.52.4
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
accelerate==1.8.1
Pillow==11.2.1
opencv-python==4.11.0.86
easyocr==1.7.2
pytesseract==0.3.13
numpy==1.26.4
pandas==2.2.3
regex==2024.11.6
fastapi
uvicorn[standard]
minio
pydantic
python-dotenv
pydantic-settings
numpy
einops
timm